reorganized half of the helper functions
