+++ server/routes/api/claimGet.js
@@ -1,6 +1,6 @@
 const { getClaim } = require('../../helpers/lbryApi.js');
 const { addGetResultsToFileData, createFileData } = require('../../helpers/publishHelpers.js');
-const { handleErrorResponse } = require('../../helpers/errorHandlers.js');
+const { handleErrorResponse } = require('../utils/errorHandlers.js');
 const db = require('../../models');
 
 /*
