+++ server/routes/api/claimResolve.js
@@ -1,5 +1,5 @@
 const { resolveUri } = require('../../helpers/lbryApi.js');
-const { handleErrorResponse } = require('../../helpers/errorHandlers.js');
+const { handleErrorResponse } = require('../utils/errorHandlers.js');
 
 /*
