+++ server/routes/api/claimShortId.js
@@ -1,4 +1,4 @@
-const { handleErrorResponse } = require('../../helpers/errorHandlers.js');
+const { handleErrorResponse } = require('../utils/errorHandlers.js');
 const db = require('../../models');
 
 /*
