+++ server/routes/assets/serveAssetByIdentifierAndClaim.js
@@ -4,8 +4,8 @@ const {
   flipClaimNameAndIdForBackwardsCompatibility,
   logRequestData,
   getClaimIdAndServeAsset,
-} = require('../../helpers/serveHelpers.js');
-const lbryUri = require('../../helpers/lbryUri.js');
+} = require('../utils/serve.js');
+const lbryUri = require('../utils/lbryUri.js');
 const handleShowRender = require('../../render/build/handleShowRender.js');
 
 const SERVE = 'SERVE';
