+++ server/routes/utils/serve.js
@@ -1,5 +1,5 @@
 const logger = require('winston');
-const { getClaimId, getLocalFileRecord } = require('../controllers/serveController.js');
+const { getClaimId, getLocalFileRecord } = require('../../controllers/serveController.js');
 const { handleErrorResponse } = require('./errorHandlers.js');
 
 const SERVE = 'SERVE';
